[ci skip] persist bloom filter, read bloom filter from disk if it exists
